+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - `balance.get()` resource implementation with tests and usage example.
+- `products.getTypes()` resource implementation with tests and usage example.
 
 ## [0.0.1] - 2026-04-07